Fix  some gameplay issues
The application used to show Lacerda's next Turn Order spots immediately
when a new card was drawn. This commit fixes that issue by adding a
Shuttle indicator, and revealing Turn Order spots only after an
additional click.

Three random numbers in the ranges 1-4, 1-6 and 1-8 are also added to
the card each Round.